backend/requirements.txt
google-cloud-speech
requests
transformers
faster-whisper
sounddevice
websockets
pydub
numpy<2
python-dotenv
sacremoses
RealtimeSTT
PyQt6